Fix handling of large ACR lists
If an ACR list was split over multiple responses,
AcrListResponse#getAcrListData used BerTlvData#getData to get the array
extra responses after the first should be copied into. However, since
BerTlvData#getData returned a copy of the array and not the actual
array, the copies had no effect.

Fix by implementing BerTlvData#addData, which copies data to the correct
position and increments the index pointer in one go, and adjust
AcrListResponse#getAcrListData to use it.

Also, remove BerTlvData#setCurrentIndex so that the atomic
BerTlvData#addData must be used instead.