In SCP03Wrapper, handle R-APDUs with error status words per spec.
Global Platform Card Specifications 2.2, Amendment D, versions 1.1.1 and higher specify in section 6.2.5:

"No R-MAC shall be generated and no protection shall be applied to a response that includes an error status word: in this case only the status word shall be returned in the response. All status words except '9000' and warning status words (i.e. '62xx' and '63xx') shall be interpreted as error status words."

This PR is an attempt to bring the handling of response APDUs in an SCP03 session in line with this part of the specification.  The previous behavior would throw an exception and terminate the session if the card returned an error in response to any command APDU inside the SCP03 RMAC session.